Compute Fairness Metrics for Binary Classification

This function evaluates a comprehensive set of fairness metrics for binary classification models across groups defined by a sensitive attribute (e.g., race, gender). It returns a unified data frame containing metric values, optionally with bootstrap confidence intervals.

Usage

get_fairness_metrics(
  data,
  outcome,
  group,
  group2 = NULL,
  condition = NULL,
  probs,
  cutoff = 0.5,
  bootstraps = 2500,
  alpha = 0.05,
  digits = 2
)

Arguments

data

A data frame containing the outcome, group, and predicted probabilities.

outcome

The name of the column containing the true binary outcome.

group

The name of the column representing the sensitive attribute (e.g., race, gender).

group2

Define if conditional statistical parity is desired. Name of a secondary group variable used for conditional fairness analysis.

condition

Define if conditional statistical parity is desired. If the conditional group is categorical, the condition supplied must be a character of the levels to condition on. If the conditional group is continuous, the conditions supplied must be a character containing the sign of the condition and the value to threshold the continuous variable (e.g. "<50", ">50", "<=50", ">=50").

probs

The name of the column with predicted probabilities.

cutoff

Numeric threshold for classification. Default is 0.5.

bootstraps

Number of bootstrap samples. Default is 2500.

alpha

Significance level for confidence intervals. Default is 0.05.

digits

Number of digits to round the metrics to. Default is 2.

confint

Logical; whether to compute bootstrap confidence intervals. Default is TRUE.

Value

A data frame with the evaluated fairness metrics.

Details

Metrics Included:

Statistical Parity: Difference in positive prediction rates across groups.

Conditional Statistical Parity (if group2 and condition are specified): Parity conditioned on a second group and value.

Equal Opportunity: Difference in true positive rates (TPR) across groups.

Predictive Equality: Difference in false positive rates (FPR) across groups.

Balance for Positive Class: Checks whether the predicted probability distributions for positive outcomes are similar across groups.

Balance for Negative Class: Same as above, but for negative outcomes.

Predictive Parity: Difference in positive predictive values (precision) across groups.

Brier Score Parity: Difference in Brier scores across groups.

Overall Accuracy Parity: Difference in overall accuracy across groups.

Treatment Equality: Ratio of false negatives to false positives across groups.

Examples

# \donttest{
library(FairnessTutorial)
library(dplyr)
library(randomForest)
data("mimic_preprocessed")
set.seed(123)
train_data <- mimic_preprocessed |>
  dplyr::filter(dplyr::row_number() <= 700)
# Fit a random forest model
rf_model <- randomForest::randomForest(factor(day_28_flg) ~ ., data = train_data, ntree = 1000)
# Test the model on the remaining data
test_data <- mimic_preprocessed |>
  dplyr::mutate(gender = ifelse(gender_num == 1, "Male", "Female"))|>
  dplyr::filter(dplyr::row_number() > 700)

test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]

# Fairness evaluation
# We will use sex as the sensitive attribute and day_28_flg as the outcome.
# We choose threshold = 0.41 so that the overall FPR is around 5%.

# Evaluate Accuracy Parity
get_fairness_metrics(
 dat = test_data,
 outcome = "day_28_flg",
 group = "gender",
 group2 = "service_unit",
 condition = "MICU",
 probs = "pred",
 cutoff = 0.5
)
#>                                                Metric GroupFemale GroupMale
#> 1                                  Statistical Parity        0.17      0.14
#> 2  Conditional Statistical Parity (MICU service_unit)        0.17      0.17
#> 3                                   Equal Opportunity        0.00      0.00
#> 4                                 Predictive Equality        0.00      0.00
#> 5                          Balance for Positive Class        0.92      0.92
#> 6                          Balance for Negative Class        0.02      0.02
#> 7                                   Predictive Parity        1.00      1.00
#> 8                                  Brier Score Parity        0.00      0.00
#> 9                             Overall Accuracy Parity        1.00      1.00
#> 10                                 Treatment Equality         NaN       NaN
#>    Difference   95% Diff CI Ratio 95% Ratio CI
#> 1        0.03 [-0.01, 0.07]  1.21 [0.91, 1.62]
#> 2        0.00 [-0.07, 0.07]  1.00 [0.65, 1.54]
#> 3        0.00        [0, 0]   NaN   [NaN, NaN]
#> 4        0.00        [0, 0]   NaN   [NaN, NaN]
#> 5        0.00 [-0.03, 0.03]  1.00 [0.97, 1.03]
#> 6        0.00 [-0.01, 0.01]  1.00  [0.48, 2.1]
#> 7        0.00        [0, 0]  1.00       [1, 1]
#> 8        0.00 [-0.01, 0.01]   NaN   [NaN, NaN]
#> 9        0.00        [0, 0]  1.00       [1, 1]
#> 10        NaN    [NaN, NaN]   NaN   [NaN, NaN]
# }