Compute Fairness Metrics for Binary Classification

DISCLAIMER: THIS DOES NOT WORK YET This function evaluates various fairness metrics on a binary classification model across different groups. The metrics are returned in a single data frame.

Usage

get_fairness_metrics(
  data,
  outcome,
  group,
  group2 = NULL,
  condition = NULL,
  probs,
  cutoff = 0.5,
  bootstraps = 2500,
  alpha = 0.05,
  digits = 2
)

Arguments

data

A data frame containing the outcome, group, and predicted probabilities.

outcome

The name of the column containing the true binary outcome.

group

The name of the column representing the sensitive attribute (e.g., race, gender).

probs

The name of the column with predicted probabilities.

cutoff

Numeric threshold for classification. Default is 0.5.

bootstraps

Number of bootstrap samples. Default is 2500.

alpha

Significance level for confidence intervals. Default is 0.05.

digits

Number of digits to round the metrics to. Default is 2.

confint

Logical; whether to compute bootstrap confidence intervals. Default is TRUE.

Value

A data frame with the evaluated fairness metrics.

Examples

# \donttest{
library(FairnessTutorial)
library(dplyr)
library(randomForest)
data("mimic_preprocessed")
set.seed(123)
train_data <- mimic_preprocessed |>
  dplyr::filter(dplyr::row_number() <= 700)
# Fit a random forest model
rf_model <- randomForest::randomForest(factor(day_28_flg) ~ ., data = train_data, ntree = 1000)
# Test the model on the remaining data
test_data <- mimic_preprocessed |>
  dplyr::mutate(gender = ifelse(gender_num == 1, "Male", "Female"))|>
  dplyr::filter(dplyr::row_number() > 700)

test_data$pred <- predict(rf_model, newdata = test_data, type = "prob")[, 2]

# Fairness evaluation
# We will use sex as the sensitive attribute and day_28_flg as the outcome.
# We choose threshold = 0.41 so that the overall FPR is around 5%.

# Evaluate Accuracy Parity
get_fairness_metrics(
  dat = test_data,
  outcome = "day_28_flg",
  group = "gender",
  probs = "pred",
  cutoff = 0.41
)
#> [[1]]
#>   Metric GroupFemale GroupMale Difference   95% Diff CI Ratio 95% Ratio CI
#> 1    PPR        0.17      0.14       0.03 [-0.01, 0.07]  1.21 [0.91, 1.62]
#> 
#> [[2]]
#>   Metric GroupFemale GroupMale Difference 95% Diff CI Ratio 95% Ratio CI
#> 1    FNR           0         0          0      [0, 0]   NaN   [NaN, NaN]
#> 
#> [[3]]
#>   Metric GroupFemale GroupMale Difference 95% Diff CI Ratio 95% Ratio CI
#> 1    FPR           0         0          0      [0, 0]   NaN   [NaN, NaN]
#> 
#> [[4]]
#>                 Metric GroupFemale GroupMale Difference   95% Diff CI Ratio
#> 1 Avg. Predicted Prob.        0.92      0.92          0 [-0.03, 0.03]     1
#>   95% Ratio CI
#> 1 [0.97, 1.03]
#> 
#> [[5]]
#>                 Metric GroupFemale GroupMale Difference   95% Diff CI Ratio
#> 1 Avg. Predicted Prob.        0.02      0.02          0 [-0.01, 0.01]     1
#>   95% Ratio CI
#> 1  [0.48, 2.1]
#> 
#> [[6]]
#>   Metric GroupFemale GroupMale Difference 95% Diff CI Ratio 95% Ratio CI
#> 1    PPV           1         1          0      [0, 0]     1       [1, 1]
#> 
#> [[7]]
#>        Metric GroupFemale GroupMale Difference   95% Diff CI Ratio 95% Ratio CI
#> 1 Brier Score           0         0          0 [-0.01, 0.01]   NaN   [NaN, NaN]
#> 
#> [[8]]
#>     Metric GroupFemale GroupMale Difference 95% Diff CI Ratio 95% Ratio CI
#> 1 Accuracy           1         1          0      [0, 0]     1       [1, 1]
#> 
#> [[9]]
#>        Metric GroupFemale GroupMale Difference 95% Diff CI Ratio 95% Ratio CI
#> 1 FN/FP Ratio         NaN       NaN        NaN  [NaN, NaN]   NaN   [NaN, NaN]
#> 
# }